one server on localhost acting as expected commit before scaling with go reverse proxy
